requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlmodel==0.0.8
pydantic>=1.8.2,<2.0.0
pytest==7.4.3
sqlalchemy>=1.4.17,<=1.4.41
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.1
limits==3.6.0
slowapi==0.1.8
python-multipart==0.0.9
